fix game results view; fixed displaying result component

+++ dots/api/models.py
@@ -6,8 +6,8 @@ class Match(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     winner = models.CharField(max_length=50)
     looser = models.CharField(max_length=50)
-    win_score = models.DecimalField(max_digits=6, decimal_places=1)
-    loose_score = models.DecimalField(max_digits=6, decimal_places=1)
+    win_score = models.IntegerField()
+    loose_score = models.IntegerField()
     equal = models.BooleanField(default=False)
 
     def create(self, validated_data):
